@theojulienne 2021-01-05
This should fix the issue with constraints that are symbols. Can anyone think of other constraint types we might be missing that we should add tests for and verify here?