@rafaelfranca 2021-01-05
Let me test in the app, but it looks good to me.